Add files via upload
sorry for the unstructured stuff. both, random selections of episodes and consistent ones are possible. I indicated it with comments, not a boolean swith or sth convenient :S
It could make sense to use some decent weights (not optimized) and then transfer those to the moving window implementation. More hyperparameters, more to tune :_: